% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{msg_conn}
\alias{msg_conn}
\title{Wrap a connection so that it can read or write msgpack objects.}
\usage{
msg_conn(conn, ...)
}
\arguments{
\item{conn}{A connection object that supports readBytes.}

\item{...}{Packing options (see pack[])}

\item{bufsize}{The size of the buffer to use to pack objects before sending. Optimise this for speed?}
}
\value{
A connection object with "class" and invisibly, the number of bytes written. NULL, invisibly.
}
\description{
Wrap a connection so that it can read or write msgpack objects.
}
\author{
Peter
}
