% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{packb}
\alias{packb}
\title{Encode R data structures in bytes using msgpack.}
\usage{
packb(datum, warn = TRUE, compatible = FALSE, use_arrays = "auto")
}
\arguments{
\item{x}{An R object.}
}
\value{
An object of class "raw".
}
\description{
Encode R data structures in bytes using msgpack.
}

