% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpackb}
\alias{unpackb}
\title{Unpack a raw byte object in msgpack format into an R data object.}
\usage{
unpackb(x)
}
\arguments{
\item{dat}{\code{\link[base:raw]{base::raw()}} byte data.}

\item{use_env}{if FALSE, msgpack dicts are unpacked into a vector
or list with a name attribute. If \code{TRUE}, msgpack dicts are
unpacked into environment objects. Note that this overrides the
option use_df.}

\item{simplify}{If \code{FALSE}, all msgpack arrays will be unpacked
into lists. If \code{TRUE}, will produce atomic vectors when possible.}

\item{transform}{A function will be called for each data element.}

\item{dict}{What type of object to produce given a msgpack dict. If
given an environment (e.g. \code{emptyenv()}, will produce an
environment with the given parent. If given a vector, will try to
produce a named list.}

\item{use_df}{If TRUE, msgpack dictionaries whose elements are all
arrays having the same length are converted to \code{\link[=data.frame]{data.frame()}}s.}

\item{nil}{If NA, will represent nil values in msgpack arrays as NA. If
\code{NULL}, will unpack nil values in arrays as \code{NULL}. If \code{c()},
will omit null values from arrays.}
}
\value{
an R data object.

The msgpack format does not have typed arrays, so all msgpack
arrays are effectively lists from the R perspective. However, if an
array containing compatibly typed elements is read, \code{unpack} will
return a logical, integer, real or string vector as
appropriate. Integer values may be converted to real, but boolean
values will not be cast to numeric, not any types to string. If
conversion from a large integer to real loses precision, a warning
is printed.

Strings are treated as UTF-8. If a msgpack string does not appear
to be a valid UTF-8 sequence, a warning is printed and a raw object
is produced instead.

Dictionary entries with keys that are other than strings are not
yet supported. If use

If \code{use_env=TRUE} and duplicate names are given in a dictionary, a
warning is produced and thew duplicate entries are dropped.
}
\description{
Unpack a raw byte object in msgpack format into an R data object.
}
