% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{unpackMsgs}
\alias{unpackMsgs}
\title{Extract a number of msgpack messages from a raw object.}
\usage{
unpackMsgs(x, n = NA, ...)
}
\arguments{
\item{x}{\code{\link[=raw]{raw()}} byte data, such as read using \code{\link[=readBin]{readBin()}}.}

\item{n}{How many messages to read. An "NA" here means to read as
much as possible.}

\item{...}{Unpacking parameters (see \code{\link[=unpackOpts]{unpackOpts()}})}
}
\value{
\link{unpackMsgs(r, n)} returns a list \code{X} with three elements:
\code{X$msgs} is a list of the messages unpacked. \code{X$remaining} is a
\code{\link[=raw]{raw()}} vector of data that was not unpacked. \code{x$status} is a
character value indicating why parsing stopped.

Some status values you many want to check for are \code{"ok"} meaning
that the requested number of messages was read; \code{"end of input"}
meaning a smaller number of messages was read; and
\code{"buffer underflow"} meaning that we were in the middle of a
message when the end of input was reached. Other status values
indicate errors encountered in parsing.
}
\description{
Extract a number of msgpack messages from a raw object.
}
