% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{msgConnection}
\alias{msgConnection}
\title{Wrap a connection so that it can read msgpack objects.}
\usage{
msgConnection(conn, read_size = NA, max_size = NA, ...)
}
\arguments{
\item{conn}{A connection object that supports readBytes.}

\item{read_size}{The default read size.}

\item{max_size}{The maximum size of a message.}

\item{...}{Packing options (see pack[])}
}
\value{
A connection object.
}
\description{
Wrap a connection so that it can read msgpack objects.
}
