% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{packb}
\alias{packb}
\title{Encode R data structures in bytes using msgpack.}
\usage{
packb(x, compatible = FALSE)
}
\arguments{
\item{x}{An R object, which can be a vectors, lists, environments,
and combinations thereof.}

\item{compatible}{If TRUE, emitted bytes conform to version 1.0 of
msgpack encoding. This means that msgpack strings are used for
raw objects.}

\item{warn}{Whether to emit warnings.}

\item{as.is}{If TRUE, vectors of length 1 having no names attribute
are encoded as msgpack arrays of length 1. Otherwise singleton
vectors are simplified to msgpack scalars.}

\item{use.dicts}{if TRUE, vectors having a "names" attribute}

\item{x}{An R object.}
}
\value{
An object of class "raw".

Strings are always emitted in UTF-8.
}
\description{
Encode R data structures in bytes using msgpack.
}
