% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpackb}
\alias{unpackb}
\title{Unpack a raw byte object in msgpack format into an R data object.}
\usage{
unpackb(dat, warn = TRUE, use_envs = FALSE, simplify = TRUE)
}
\arguments{
\item{dat}{A raw byte object.}

\item{warn}{Whether to generate warnings.}

\item{simplify}{If `FALSE`, arrays are always represented as lists.}

\item{dicts_as_environments}{Whether to represent msgpack dicts as
R environments.}
}
\value{
an R data object.
}
\description{
Unpack a raw byte object in msgpack format into an R data object.
}

